+++ crates/blockifier/src/transaction/account_transactions_test.rs
@@ -22,8 +22,9 @@ use crate::fee::fee_utils::{calculate_tx_l1_gas_usage, get_fee_by_l1_gas_usage};
 use crate::fee::gas_usage::estimate_minimal_l1_gas;
 use crate::invoke_tx_args;
 use crate::state::state_api::{State, StateReader};
+use crate::test_utils::invoke::InvokeTxArgs;
 use crate::test_utils::{
-    create_calldata, declare_tx, deploy_account_tx, InvokeTxArgs, NonceManager, BALANCE,
+    create_calldata, declare_tx, deploy_account_tx, NonceManager, BALANCE,
     DEFAULT_STRK_L1_GAS_PRICE, GRINDY_ACCOUNT_CONTRACT_CAIRO0_PATH, MAX_FEE, MAX_L1_GAS_AMOUNT,
     MAX_L1_GAS_PRICE, TEST_ACCOUNT_CONTRACT_CLASS_HASH, TEST_CONTRACT_ADDRESS,
     TEST_FAULTY_ACCOUNT_CONTRACT_ADDRESS, TEST_GRINDY_ACCOUNT_CONTRACT_CLASS_HASH,
@@ -436,8 +437,11 @@ fn test_recursion_depth_exceeded(
             stark_felt!(exceeding_recursion_depth),
         ],
     );
-    let invoke_args =
-        InvokeTxArgs { calldata, nonce: nonce_manager.next(account_address), ..invoke_args };
+    let invoke_args = crate::test_utils::invoke::InvokeTxArgs {
+        calldata,
+        nonce: nonce_manager.next(account_address),
+        ..invoke_args
+    };
     let tx_execution_info = run_invoke_tx(&mut state, &block_context, invoke_args);
 
     assert!(tx_execution_info.unwrap().revert_error.unwrap().contains("recursion depth exceeded"));
